seras-test-generator/pyproject.toml
[project]
name = "seras-test-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.8",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.40.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pytest.PytestCollectionWarning",
]
